Modify the generated annotation rakefile based on Rails version
`rake routes` was marked for deprecation as of Rails 6.0.0, and will be fully deprecated by Rails 6.1.0. We, on the other hand, want to continue annotating our routes.

Thus, we provide an option by which we can set the routes-lister command.

This assumes we have Rails installed, but since we are already checking the Rails environment, we are not introducing any sort of new dependency.